+++ cypress/calculator.spec.js
@@ -1,3 +1,3 @@
 it("2개의 숫자에 대해 덧셈이 가능하다.", () => {
   cy.visit("index.html");
-}) 아아아ㅏㅏㅏㅏㅏ
+});
