+++ docs/README.md
@@ -4,13 +4,12 @@
 
 2. 숫자 버튼 클릭 이벤트
 
-2-1. - [ ] 입력값이 비어있으면 안된다.
+- 입력값이 비어있으면 안된다.
+- 입력값은 숫자 타입이어야 한다.
+- 입력값이 3개 이상 들어오면 에러 처리
+- 입력값이 4자리 이상이면 에러 처리
 
-- [ ] 입력값은 숫자 타입이어야 한다.
-- [ ] 입력값이 3개 이상 들어오면 에러 처리
-- [ ] 입력값이 4자리 이상이면 에러 처리
-
-2-2.data-type로 숫자를 추출하여, 결과창 뒤에 붙인다.
+2-2. data-type로 숫자를 추출하여, 결과창 뒤에 붙인다.
 
 3.  +-\*/ 버튼 클릭 이벤트
 
@@ -20,7 +19,7 @@
 
 3-3. 연산자 입력이 연속될 수 없다.
 
-3-2. 연산자 버튼을 누르면, 해당 연산자를 결과창 뒤에 붙인다.
+3-4. 연산자 버튼을 누르면, 해당 연산자를 결과창 뒤에 붙인다.
 
 4. = 버튼
