+++ src/js/constants.js
@@ -1,0 +1,17 @@
+const MSG = {
+  exceed: '글자 수를 초과했습니다!',
+  empty_number: '숫자를 먼저 입력해주세요!',
+  lack_number: '숫자 입력이 부족합니다!',
+  have_operator: '이미 연산자를 입력하셨습니다.',
+};
+
+const MAX_DIGIT = 3;
+const ZERO = '0';
+const EMPTY_STRING = '';
+
+export {
+  MSG,
+  MAX_DIGIT,
+  ZERO,
+  EMPTY_STRING
+}
