+++ src/js/utils.js
@@ -1,0 +1,7 @@
+const $ = (selector) => document.querySelector(selector);
+const $$ = (selector) => document.querySelectorAll(selector);
+
+export {
+  $,
+  $$
+}
